requirements.txt
streamlit==1.30.0
pillow==10.1.0
rembg==2.0.0
onnxruntime==1.22.1